refactor(search): split match.go — extract MatchFrom into per-family helpers
match.go was the largest non-test file (1516 lines) and MatchFrom alone
was a ~940-line function that had grown one inline projection block per
content family — every new family meant editing the monolith.

match.go now holds only the Match wire struct. The projection moves to a
new match_from.go: a slim MatchFrom orchestrator plus 20 small
applyXxxAttrs(m *Match, a *celexpr.FileAttributes) helpers, one per
family (image / media / archive+binary / email+source / frontmatter /
disk-image / install-package / codesign / plist / bytecode / science /
database / bookmark / chat / font / xattr / …). Adding a content family
now touches one focused helper + one orchestrator call.

Pure mechanical extraction — helper bodies are verbatim, run in the same
order, and the orchestrator preserves the nil-Attrs / nil-Extra
short-circuits. Verified behaviour-identical: JSON output over the 65
fixture files spanning 57 content types (with --with-hashes
--check-disguised --with-xattrs) is byte-for-byte unchanged (same
sha256). go vet / golangci-lint / full -race suite all green.
